backend/requirements.txt
# PACO Backend - Python Dependencies

# Web Framework
fastapi>=0.110.0
uvicorn[standard]>=0.27.1
python-multipart>=0.0.9

# Database
sqlalchemy>=2.0.25
alembic>=1.13.1
asyncpg>=0.30.0
psycopg2-binary>=2.9.9
greenlet>=3.0.0

# Redis
redis>=5.0.1

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
pydantic[email]>=2.6.1
pydantic-settings>=2.1.0

# HTTP Client
httpx>=0.26.0
aiohttp>=3.9.3

# OpenTelemetry (optional for now)
# opentelemetry-api>=1.22.0
# opentelemetry-sdk>=1.22.0
# opentelemetry-exporter-otlp>=1.22.0
# opentelemetry-instrumentation-fastapi>=0.43b0

# Langfuse Client
langfuse>=2.0.1

# YAML Processing
pyyaml>=6.0.1

# Cron expression parsing (company heartbeat scheduler)
croniter>=2.0.0

# Utilities
python-dateutil>=2.8.2
orjson>=3.9.12

# Template Engine (for code generation)
Jinja2>=3.1.0

# Docker SDK (for infrastructure deployment)
docker>=7.0.0

# WebSocket support (for log streaming)
websockets>=12.0

# PM2 Process Management (via shell)
# No Python package needed - uses subprocess

# Development
pytest>=7.0.0
pytest-asyncio>=0.23.4
black>=24.1.1
ruff>=0.1.14
